+++ src/main/java/com/inq/wishhair/wesharewishhair/review/controller/LikeReviewController.java
@@ -1,0 +1,27 @@
+package com.inq.wishhair.wesharewishhair.review.controller;
+
+import com.inq.wishhair.wesharewishhair.auth.config.resolver.ExtractPayload;
+import com.inq.wishhair.wesharewishhair.review.service.LikeReviewService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class LikeReviewController {
+
+    private final LikeReviewService likeReviewService;
+
+    @PostMapping("/review/like/{reviewId}")
+    public ResponseEntity<Void> likeReview(
+            @PathVariable Long reviewId,
+            @ExtractPayload Long userId) {
+
+        likeReviewService.LikeReview(reviewId, userId);
+        return ResponseEntity.noContent().build();
+    }
+}
