+++ src/main/java/com/inq/wishhair/wesharewishhair/review/service/ReviewService.java
@@ -31,7 +31,6 @@ public class ReviewService {
     private final UserRepository userRepository;
     private final HairStyleRepository hairStyleRepository;
     private final PhotoStore photoStore;
-    private final LikeReviewRepository likeReviewRepository;
 
     @Transactional
     public Long createReview(ReviewCreateDto dto) {
@@ -53,20 +52,6 @@ public class ReviewService {
         return reviewRepository.save(review).getId();
     }
 
-    @Transactional
-    public void LikeReview(Long reviewId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new WishHairException(ErrorCode.NOT_EXIST_KEY));
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new WishHairException(ErrorCode.NOT_EXIST_KEY));
-
-        likeReviewRepository.findByUserAndReview(user, review)
-                .ifPresentOrElse(likeReviewRepository::delete, () -> {
-                    LikeReview likeReview = LikeReview.createLikeReview(user, review);
-                    likeReviewRepository.save(likeReview);
-                });
-    }
-
     public List<ReviewResponse> getReviews(Pageable pageable, String condition) {
         List<Review> reviews = reviewRepository.findReviewByPaging(pageable);
         // Query 에서 정렬이 안돼서 Service 에서 정렬
@@ -77,7 +62,7 @@ public class ReviewService {
         return toResponse(reviews);
     }
 
-    public List<ReviewResponse> toResponse(List<Review> reviews) {
+    private List<ReviewResponse> toResponse(List<Review> reviews) {
         return reviews.stream()
                 .map(ReviewResponse::new)
                 .toList();
